+++ src/elements/NestedElementManager.php
@@ -696,7 +696,8 @@ JS, [
             // If this is a draft, its nested element ownership will be duplicated by Drafts::createDraft()
             if ($owner->getIsRevision()) {
                 $this->createRevisions($owner->duplicateOf, $owner);
-            } elseif ($owner->getIsCanonical()) {
+            // getIsUnpublishedDraft is needed for "save as new" duplication
+            } elseif (!$owner->getIsDraft() || $owner->getIsUnpublishedDraft()) {
                 $this->duplicateNestedElements($owner->duplicateOf, $owner, true, !$isNew);
             }
             $resetValue = true;
@@ -1113,7 +1114,9 @@ JS, [
                             'sortOrder' => $element->getSortOrder(),
                         ], updateTimestamp: false);
                     } else {
-                        $newElementId = $element->getCanonicalId();
+                        // If we're updating the canonical owner element, then go with the nested element’s canonical ID.
+                        // Otherwise, leave the current ID intact.
+                        $newElementId = $target->getIsCanonical() ? $element->getCanonicalId() : $element->id;
                     }
                 } elseif (!$force && $element->getPrimaryOwnerId() === $target->id) {
                     // Only the element ownership was duplicated, so just update its sort order for the target element